Moved management of additional `metadata_fields` from the `SearchGet` service to the `DefaultJSONSummarySerializer` created for that, it will override the default `plone.restapi` `DefaultJSONSummarySerializer` and add `id` and `UID` by default to the results.
