+++ src/imio/restapi/services/search.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from imio.helpers.content import uuidsToObjects
-from imio.restapi.utils import listify
 from plone.app.querystring.queryparser import parseFormquery
 from plone.restapi.search.handler import SearchHandler
 from plone.restapi.search.utils import unflatten_dotted_dict
@@ -11,13 +10,7 @@ from plone.restapi.services.search.get import SearchGet as BaseSearchGet
 class SearchGet(BaseSearchGet):
     """Base SearchGet, handles :
        - using a base_search_uid (Collection UID) as base query;
-       - adding additional parameters to query;
-       - formalize metadata_fields."""
-
-    @property
-    def _additional_fields(self):
-        """By default add 'UID' to returned data."""
-        return ["UID"]
+       - adding additional parameters to query."""
 
     def _set_query_before_hook(self):
         """Manipulate query before hook."""
@@ -51,14 +44,6 @@ class SearchGet(BaseSearchGet):
            WARNING plone.restapi.search.query No such index: 'my_custom_parameter'"""
         query.pop("my_custom_parameter", None)
 
-    def _set_metadata_fields(self):
-        """Must be set in request.form."""
-        form = self.request.form
-        # manage metadata_fields
-        additional_metadata_fields = listify(form.get("metadata_fields", []))
-        additional_metadata_fields += self._additional_fields
-        form["metadata_fields"] = additional_metadata_fields
-
     def _process_reply(self):
         """Easier to override if necessary to call various ways from reply method."""
         query = {}
@@ -70,8 +55,6 @@ class SearchGet(BaseSearchGet):
         query.update(self._set_query_after_hook())
         self._clean_query(query)
         query = unflatten_dotted_dict(query)
-
-        self._set_metadata_fields()
 
         return SearchHandler(self.context, self.request).search(query)
